requirements.txt
# Core dependencies
fastapi==0.110.3
uvicorn==0.29.0

# CrewAI (keep versions pinned for stability)
crewai==0.130.0
crewai-tools==0.47.1

# LangChain & OpenAI for LLM integration
langchain-core>=0.1.52
langchain-openai>=0.2.1,<0.4.0
openai>=1.68.2,<2.0.0

# PDF and data processing
pypdf>=4.2.0
pandas>=2.2.2
numpy>=1.26.4

# Utils
python-dotenv>=1.0.1
jinja2>=3.1.4
